back-end/src/main/java/com/example/bookstore/controllers/UserController.java
package com.example.bookstore.controllers;

import com.example.bookstore.dtos.UserDto;
import com.example.bookstore.mappers.UserMapper;
import com.example.bookstore.services.IUserService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import jakarta.validation.Valid;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/v1/users")
@Slf4j
public class UserController {

    private IUserService service;

    private UserMapper mapper;

    @Autowired
    public UserController(IUserService service, UserMapper mapper) {
        this.service = service;
        this.mapper = mapper;
    }

    @PostMapping("/register")
    @ResponseStatus(HttpStatus.CREATED)
    public UserDto register(@Valid @RequestBody UserDto user) {
        log.info("register({})", user.getUsername());
        return mapper.toDto(service.createUser(mapper.toEntity(user)));
    }

    @Operation(
            requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(
                    required = true,
                    content = @Content(
                            mediaType = "application/json",
                            schema = @Schema(example = "{ \"username\": \"string\", \"password\": \"string\" }")
                    )
            ),
            responses = {
                    @ApiResponse(responseCode = "200", description = "Login successful"),
                    @ApiResponse(responseCode = "401", description = "Invalid credentials")
            }
    )
    @PostMapping("/login")
    public void loginDocumentation() {
        // For OpenAPI documentation only.
        return;
    }

}
